Move `propose_state` from `OrionClient` to `engine`
This method is performing more engine logic and does not map to a single API route. It makes increasing sense for this logic to take place in the engine alongside other client-side orchestration logic. This change should reduce circular imports while addressing result persistence and tracking during state proposal.
