+++ tests/test_engine.py
@@ -84,29 +84,6 @@ def flow_run_caplog(caplog):
 
 
 @pytest.fixture
-def mock_client_sleep(monkeypatch):
-    """
-    Mock sleep used by the orion_client to not actually sleep but to set the
-    current time to now + sleep delay seconds.
-    """
-    original_now = pendulum.now
-    time_shift = 0
-
-    async def callback(delay_in_seconds):
-        nonlocal time_shift
-        time_shift += delay_in_seconds
-
-    monkeypatch.setattr(
-        "pendulum.now", lambda *args: original_now(*args).add(seconds=time_shift)
-    )
-
-    sleep = AsyncMock(side_effect=callback)
-    monkeypatch.setattr("prefect.client.sleep", sleep)
-
-    return sleep
-
-
-@pytest.fixture
 async def flow_run_context(orion_client, local_filesystem):
     @flow
     def foo():
@@ -131,7 +108,7 @@ class TestOrchestrateTaskRun:
         self,
         orion_client,
         flow_run,
-        mock_client_sleep,
+        mock_anyio_sleep,
         local_filesystem,
         monkeypatch,
     ):
@@ -161,12 +138,12 @@ class TestOrchestrateTaskRun:
             client=orion_client,
         )
 
-        mock_client_sleep.assert_awaited_once()
+        mock_anyio_sleep.assert_awaited_once()
         assert state.is_completed()
         assert state.result() == 1
 
     async def test_does_not_wait_for_scheduled_time_in_past(
-        self, orion_client, flow_run, mock_client_sleep, local_filesystem
+        self, orion_client, flow_run, mock_anyio_sleep, local_filesystem
     ):
         @task
         def foo():
@@ -194,12 +171,12 @@ class TestOrchestrateTaskRun:
             client=orion_client,
         )
 
-        mock_client_sleep.assert_not_called()
+        mock_anyio_sleep.assert_not_called()
         assert state.is_completed()
         assert state.result() == 1
 
     async def test_waits_for_awaiting_retry_scheduled_time(
-        self, mock_client_sleep, orion_client, flow_run, local_filesystem
+        self, mock_anyio_sleep, orion_client, flow_run, local_filesystem
     ):
         # Define a task that fails once and then succeeds
         mock = MagicMock()
@@ -238,8 +215,8 @@ class TestOrchestrateTaskRun:
         # Assert that the sleep was called
         # due to network time and rounding, the expected sleep time will be less than
         # 43 seconds so we test a window
-        mock_client_sleep.assert_awaited_once()
-        assert 40 < mock_client_sleep.call_args[0][0] < 43
+        mock_anyio_sleep.assert_awaited_once()
+        assert 40 < mock_anyio_sleep.call_args[0][0] < 43
 
         # Check expected state transitions
         states = await orion_client.read_task_run_states(task_run.id)
@@ -432,7 +409,7 @@ class TestOrchestrateFlowRun:
         )
 
     async def test_waits_until_scheduled_start_time(
-        self, orion_client, mock_client_sleep, partial_flow_run_context
+        self, orion_client, mock_anyio_sleep, partial_flow_run_context
     ):
         @flow
         def foo():
@@ -457,11 +434,11 @@ class TestOrchestrateFlowRun:
             partial_flow_run_context=partial_flow_run_context,
         )
 
-        mock_client_sleep.assert_awaited_once()
+        mock_anyio_sleep.assert_awaited_once()
         assert state.result() == 1
 
     async def test_does_not_wait_for_scheduled_time_in_past(
-        self, orion_client, mock_client_sleep, partial_flow_run_context
+        self, orion_client, mock_anyio_sleep, partial_flow_run_context
     ):
         @flow
         def foo():
@@ -487,11 +464,11 @@ class TestOrchestrateFlowRun:
                 partial_flow_run_context=partial_flow_run_context,
             )
 
-        mock_client_sleep.assert_not_called()
+        mock_anyio_sleep.assert_not_called()
         assert state.result() == 1
 
     async def test_waits_for_awaiting_retry_scheduled_time(
-        self, orion_client, mock_client_sleep, partial_flow_run_context
+        self, orion_client, mock_anyio_sleep, partial_flow_run_context
     ):
         flow_run_count = 0
 
@@ -524,8 +501,8 @@ class TestOrchestrateFlowRun:
         # Assert that the sleep was called
         # due to network time and rounding, the expected sleep time will be less than
         # 43 seconds so we test a window
-        mock_client_sleep.assert_awaited_once()
-        assert 40 < mock_client_sleep.call_args[0][0] < 43
+        mock_anyio_sleep.assert_awaited_once()
+        assert 40 < mock_anyio_sleep.call_args[0][0] < 43
 
         # Check expected state transitions
         states = await orion_client.read_flow_run_states(flow_run.id)
@@ -661,7 +638,7 @@ class TestFlowRunCrashes:
         self, flow_run, orion_client, monkeypatch, interrupt_type
     ):
         monkeypatch.setattr(
-            "prefect.client.OrionClient.propose_state",
+            "prefect.engine.propose_state",
             MagicMock(side_effect=interrupt_type()),
         )
